+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - VFS Encryption
-- File viewer in chats (Image/Animated Image/Video/Texts) with the option to export the file.
+- File viewer in chats (Image/Animated Image/Video/Texts/Pdf) with the option to export the file.
 - Accept/decline CLI commands.
 - Colored Emojis with its own font family.
 - OAuth2 connection to retrieve remote provisioning (Experimental and not usable without configuration).